requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
flask
huggingface-hub==0.19.4
python-dotenv
pypdf
langchain-pinecone
llama-cpp-python
langchain_community
wandb
chromadb


-e .



